Refs #16986 - Add breadcrumbs, improve layout
