+++ app/controllers/template_syncs_controller.rb
@@ -1,35 +1,4 @@
-require 'ostruct'
-
-class TemplateSyncsController < ApplicationController
-  include ::Foreman::Controller::Parameters::TemplateParams
-
+class TemplateSyncsController < ReactController
   def index
-  end
-
-  def sync_settings
-    import_settings = Setting.where :name => Setting::TemplateSync.import_setting_names
-    export_settings = Setting.where :name => Setting::TemplateSync.export_setting_names
-    @results = OpenStruct.new(:import => import_settings, :export => export_settings)
-  end
-
-  def import
-    @parse_result = OpenStruct.new ForemanTemplates::TemplateImporter.new(ui_template_import_params).import!
-  end
-
-  def export
-    @result = ForemanTemplates::TemplateExporter.new(ui_template_export_params).export!
-  end
-
-  def action_permission
-    case params[:action]
-    when 'sync_settings'
-      :view
-    else
-      super
-    end
-  end
-
-  def parameter_filter_context
-    Foreman::ParameterFilter::Context.new(:api, controller_name, params[:action])
   end
 end
