--- app/views/template_syncs/template_import_results.json.rabl
@@ -1,3 +1,0 @@
-collection @templates
-
-extends "template_syncs/template_import_result"
