+++ app/views/ui_template_syncs/export.json.rabl
@@ -5,5 +5,5 @@ attributes :repo, :branch, :git_user, :error, :warning
 node(:result_action) { 'export' }
 
 child :templates => :templates do
-  extends 'template_syncs/template_export_results'
+  extends 'ui_template_syncs/template_export_results'
 end
