+++ app/views/ui_template_syncs/import.json.rabl
@@ -5,5 +5,5 @@ attributes :repo, :branch
 node(:result_action) { 'import' }
 
 child :results => :templates do
-  extends 'template_syncs/template_import_results'
+  extends 'ui_template_syncs/template_import_results'
 end
