+++ app/views/ui_template_syncs/sync_settings.json.rabl
