+++ app/views/ui_template_syncs/template_attrs.json.rabl
