+++ app/views/ui_template_syncs/template_export_result.rabl
@@ -3,5 +3,5 @@ object @template
 attributes :name
 
 node(false) do |template|
-  partial "template_syncs/template_attrs", :object => template
+  partial "ui_template_syncs/template_attrs", :object => template
 end
