+++ app/views/ui_template_syncs/template_import_result.json.rabl
@@ -3,5 +3,5 @@ object @template_result
 attributes :additional_errors, :errors, :exception_message, :name, :template_file
 
 node(false) do |result|
-  partial "template_syncs/template_attrs", :object => result.template
+  partial "ui_template_syncs/template_attrs", :object => result.template
 end
