+++ app/views/ui_template_syncs/template_import_results.json.rabl
@@ -1,0 +1,3 @@
+collection @templates
+
+extends "ui_template_syncs/template_import_result"
